test(orchestration): move the send fleet-echo cases into their own file
orchestration-send.test.ts is at 790 counted lines on main; appending the
five fleet-echo cases pushed it to 844 and over the 800-line budget. Same
split as db-dispatch-liveness.test.ts rather than a suppression.
